+++ packages/cli/src/util/projects/link.ts
@@ -47,6 +47,10 @@ const linkSchema = {
       type: 'string',
       minLength: 1,
     },
+    projectName: {
+      type: 'string',
+      minLength: 1,
+    },
   },
 };
 
@@ -245,7 +249,7 @@ export async function getLinkedProject(
 
       if (err.missingToken || err.invalidToken) {
         throw new InvalidToken();
-      } else {
+      } else if (err.code === 'forbidden' || err.code === 'team_unauthorized') {
         throw new NowBuildError({
           message: `Could not retrieve Project Settings. To link your Project, remove the ${outputCode(
             VERCEL_DIR
@@ -318,7 +322,10 @@ export async function linkFolderToProject(
 
   await writeFile(
     join(path, VERCEL_DIR, VERCEL_DIR_PROJECT),
-    JSON.stringify(projectLink)
+    JSON.stringify({
+      ...projectLink,
+      projectName,
+    })
   );
 
   await writeReadme(path);