+++ packages/functions/src/index.ts
@@ -4,4 +4,8 @@ export { getEnv } from './get-env';
 export { waitUntil } from './wait-until';
 export { rewrite, next } from './middleware';
 export { getCache } from './cache';
+export {
+  attachDatabasePool,
+  experimental_attachDatabasePool,
+} from './db-connections';
 export type { RuntimeCache } from './cache/types';