+++ packages/hydrogen/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/hydrogen
 
+## 1.2.3
+
+### Patch Changes
+
+- Reverting support for `preferredRegion` ([#13566](https://github.com/vercel/vercel/pull/13566))
+
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies [[`6c8e763ab63c79e12c7d5455fd79cf158f43cc77`](https://github.com/vercel/vercel/commit/6c8e763ab63c79e12c7d5455fd79cf158f43cc77)]:
+  - @vercel/static-config@3.1.1
+
 ## 1.2.1
 
 ### Patch Changes